services/trainer/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
ta
sqlalchemy
psycopg2-binary
kafka-python
python-dateutil
